[Client] Add embed for editor
